Checklist feature prototype.
This is an early version of the checklist feature that I came up with. After Tabitha and I started working together, things changed and the checklist doesn't really look like this anymore. With that said, the features that have endured from this are the color coded notifaction system and the step-list.